Move directory creation into Utilities
